refactor Onboarding2RolesPanel
